Move labeling.* to src/alg/ and Graphcut now uses float weights.

+++ src/alg/labeling.hpp
@@ -29,7 +29,7 @@ struct ForSmoothFn
 };
 
 inline int label_dist(int value);
-int smoothFn(int p1, int p2, int l1, int l2, void *data);
+float smoothFn(int p1, int p2, int l1, int l2, void *data);
 
 } /* namespace elib */
